requirements.txt
fastapi
uvicorn
pydantic
requests
opencc-python-reimplemented
python-dotenv